requirements.txt
python-dotenv==1.0.1
langchain==0.3.20
langchain[groq]==0.3.20
streamlit==1.43.2
langchain_community==0.3.19
psycopg2==2.9.10
langchain_huggingface==0.1.2
unstructured[md]==0.17.0
smolagents==1.12.0
pandas==2.2.3
mcp==1.9.1
mcp[cli]==1.9.1
smolagents[mcp]==1.12.0